fix(graphql): clean code & improve UI
